+++ lib/runner/cli/cli.js
@@ -1,8 +1,7 @@
 const path = require('path');
 const lodashClone = require('lodash.clone');
-const util = require('util');
-const BStackLocal = require('browserstack-local').Local;
 const ArgvSetup = require('./argv-setup.js');
+const LocalTunnel = require('./local-tunnel.js');
 const Settings = require('../../settings/settings.js');
 const Globals = require('../../testsuite/globals.js');
 const Concurrency = require('../concurrency');
@@ -11,52 +10,6 @@ const Runner = require('../runner.js');
 const ProcessListener = require('../process-listener.js');
 const analyticsCollector = require('../../utils/analytics.js');
 const {Logger, singleSourceFile} = Utils;
-
-class LocalTunnel {
-  constructor(settings = {}) {
-    this.settings = settings;
-    this.tunnel = this.settings.tunnel;
-    this.opts = this.tunnel ? this.tunnel.options : {};
-    this.key = this.opts.key;
-    if(this.tunnel && this.tunnel.type) {
-      switch(this.tunnel.type) {
-        case 'browserstack':
-          this.local = new BStackLocal();
-      }
-    }
-  }
-
-  async start() {
-    if (this.local && this.key) {
-      try {
-        await util.promisify(this.local.start.bind(this.local))({ key: this.key, ...this.opts });
-        this.localStarted = true;
-        this.settings.local_setup = {
-          localIdentifier: this.opts.localIdentifier
-        }
-        let signal = 0;
-        const handler = async () => {
-          signal++;
-          signal === 1 && await this.stop();
-          process.exit();
-        }
-        process.on('SIGINT', handler);
-        process.on('SIGTERM', handler);
-      } catch (err) {
-        Logger.error("Starting of local failed", err);
-      }
-    }
-  }
-
-  async stop() {
-    try {
-      this.localStarted && await util.promisify(this.local.stop.bind(this.local))();
-    } catch (err) {
-      Logger.error("Stopping of local failed", err);
-    }
-  }
-};
-
 
 class CliRunner {
   static get CONFIG_FILE_JS() {
@@ -211,7 +164,6 @@ class CliRunner {
   async setupAsync(settings) {
     this.baseSettings = await this.loadConfig();
     this.commonSetup(settings);
-    this.localTunnel = new LocalTunnel(this.test_settings);
 
     return this;
   }
@@ -236,7 +188,6 @@ class CliRunner {
   setup(settings) {
     this.baseSettings = this.loadConfig();
     this.commonSetup(settings);
-    this.localTunnel = new LocalTunnel(this.test_settings);
 
     return this;
   }
@@ -262,6 +213,7 @@ class CliRunner {
       test_runner: this.test_settings.test_runner ? this.test_settings.test_runner.type : null
     });
 
+    this.localTunnel = new LocalTunnel(this.test_settings);
 
     return this;
   }
