+++ lib/transport/selenium-webdriver/browserstack.js
@@ -70,10 +70,10 @@ class Browserstack extends BaseDriver {
       this.settings.desiredCapabilities['bstack:options'].buildName = 'nightwatch-test-build';
     }
 
-    if(this.settings.local_setup) {
+    if(this.settings.browserstackLocal) {
       this.settings.desiredCapabilities['bstack:options'].local = true;
-      (this.settings.local_setup.localIdentifier && this.settings.local_setup.localIdentifier != "undefined") && 
-        (this.settings.desiredCapabilities['bstack:options'].localIdentifier = this.settings.local_setup.localIdentifier);
+      (this.settings.browserstackLocal.localIdentifier && this.settings.browserstackLocal.localIdentifier != "undefined") && 
+        (this.settings.desiredCapabilities['bstack:options'].localIdentifier = this.settings.browserstackLocal.localIdentifier);
     }
 
     if (this.local) {
